Split EllipseTool, RectangleTool, RectangleToolEx
