+++ srcjs/ellipsetool.js
@@ -1,0 +1,90 @@
+import { svg } from 'lit-element';
+import { classMap } from 'lit-html/directives/class-map.js';
+import { styleMap } from 'lit-html/directives/style-map.js';
+
+import Merge from './merge';
+import Utils from './utils';
+import BaseTool from './basetool';
+
+/** ****************************************************************************
+  * EllipseTool class
+  *
+  * Summary.
+  *
+  */
+
+export default class EllipseTool extends BaseTool {
+  constructor(argToolset, argConfig, argPos) {
+    const DEFAULT_ELLIPSE_CONFIG = {
+      position: {
+        cx: 50,
+        cy: 50,
+        radiusx: 50,
+        radiusy: 25,
+      },
+      classes: {
+        tool: {
+          'sak-ellipse': true,
+          hover: true,
+        },
+        ellipse: {
+          'sak-ellipse__ellipse': true,
+        },
+      },
+      styles: {
+        ellipse: {
+        },
+      },
+    };
+
+    super(argToolset, Merge.mergeDeep(DEFAULT_ELLIPSE_CONFIG, argConfig), argPos);
+
+    this.svg.radiusx = Utils.calculateSvgDimension(argConfig.position.radiusx);
+    this.svg.radiusy = Utils.calculateSvgDimension(argConfig.position.radiusy);
+
+    this.classes.ellipse = {};
+    this.styles.ellipse = {};
+
+    if (this.dev.debug) console.log('EllipseTool constructor coords, dimensions', this.coords, this.dimensions, this.svg, this.config);
+  }
+
+  /** *****************************************************************************
+  * EllipseTool::_renderEllipse()
+  *
+  * Summary.
+  * Renders the ellipse using precalculated coordinates and dimensions.
+  * Only the runtime style is calculated before rendering the ellipse
+  *
+  */
+
+  _renderEllipse() {
+    this.MergeAnimationClassIfChanged();
+    this.MergeAnimationStyleIfChanged();
+    this.MergeColorFromState(this.styles.ellipse);
+
+    if (this.dev.debug) console.log('EllipseTool - renderEllipse', this.svg.cx, this.svg.cy, this.svg.radiusx, this.svg.radiusy);
+
+    return svg`
+      <ellipse class="${classMap(this.classes.ellipse)}"
+        cx="${this.svg.cx}"% cy="${this.svg.cy}"%
+        rx="${this.svg.radiusx}" ry="${this.svg.radiusy}"
+        style="${styleMap(this.styles.ellipse)}"/>
+      `;
+  }
+
+  /** *****************************************************************************
+  * EllipseTool::render()
+  *
+  * Summary.
+  * The render() function for this object.
+  *
+  */
+  render() {
+    return svg`
+      <g id="ellipse-${this.toolId}" class="${classMap(this.classes.tool)}"
+        @click=${(e) => this.handleTapEvent(e, this.config)}>
+        ${this._renderEllipse()}
+      </g>
+    `;
+  }
+} // END of class
