+++ srcjs/swiss-army-knife-card.js
@@ -65,6 +65,9 @@ import LineTool from './linetool';
 import SwitchTool from './switchtool';
 import RegPolyTool from './regpolytool';
 import UserSvgTool from './usersvg';
+import EllipseTool from './ellipsetool';
+import RectangleTool from './rectangletool';
+import RectangleToolEx from './rectangletoolex';
 
 // Original injector is buggy. Use a patched version, and store this local...
 import * as SvgInjector from '../dist/SVGInjector.min.js'; // lgtm[js/unused-local-variable]
@@ -450,327 +453,15 @@ class Toolset {
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  */
 
-/** ****************************************************************************
-  * RectangleTool class
-  *
-  * Summary.
-  *
-  */
-
-class RectangleTool extends BaseTool {
-  constructor(argToolset, argConfig, argPos) {
-    const DEFAULT_RECTANGLE_CONFIG = {
-      position: {
-        cx: 50,
-        cy: 50,
-        width: 50,
-        height: 50,
-        rx: 0,
-      },
-      classes: {
-        tool: {
-          'sak-rectangle': true,
-          hover: true,
-        },
-        rectangle: {
-          'sak-rectangle__rectangle': true,
-        },
-      },
-      styles: {
-        rectangle: {
-        },
-      },
-    };
-
-    super(argToolset, Merge.mergeDeep(DEFAULT_RECTANGLE_CONFIG, argConfig), argPos);
-    this.svg.rx = argConfig.position.rx ? Utils.calculateSvgDimension(argConfig.position.rx) : 0;
-
-    this.classes.rectangle = {};
-    this.styles.rectangle = {};
-
-    if (this.dev.debug) console.log('RectangleTool constructor config, svg', this.toolId, this.config, this.svg);
-  }
-
-  /** *****************************************************************************
-  * RectangleTool::value()
-  *
-  * Summary.
-  * Receive new state data for the entity this rectangle is linked to. Called from set hass;
-  *
-  */
-  set value(state) {
-    const changed = super.value = state;
-
-    return changed;
-  }
-
-  /** *****************************************************************************
-  * RectangleTool::_renderRectangle()
-  *
-  * Summary.
-  * Renders the circle using precalculated coordinates and dimensions.
-  * Only the runtime style is calculated before rendering the circle
-  *
-  */
-
-  _renderRectangle() {
-    this.MergeAnimationClassIfChanged();
-    this.MergeAnimationStyleIfChanged();
-    this.MergeColorFromState(this.styles.rectangle);
-
-    return svg`
-      <rect class="${classMap(this.classes.rectangle)}"
-        x="${this.svg.x}" y="${this.svg.y}" width="${this.svg.width}" height="${this.svg.height}" rx="${this.svg.rx}"
-        style="${styleMap(this.styles.rectangle)}"/>
-      `;
-  }
-
-  /** *****************************************************************************
-  * RectangleTool::render()
-  *
-  * Summary.
-  * The render() function for this object.
-  *
-  */
-  render() {
-    return svg`
-      <g id="rectangle-${this.toolId}" class="${classMap(this.classes.tool)}" transform-origin="${this.svg.cx}px ${this.svg.cy}px"
-        @click=${(e) => this.handleTapEvent(e, this.config)}>
-        ${this._renderRectangle()}
-      </g>
-    `;
-  }
-} // END of class
-
 /**
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  */
 
-/** ****************************************************************************
-  * RectangleToolEx class
-  *
-  * Summary.
-  *
-  */
-
-class RectangleToolEx extends BaseTool {
-  constructor(argToolset, argConfig, argPos) {
-    const DEFAULT_RECTANGLEEX_CONFIG = {
-      position: {
-        cx: 50,
-        cy: 50,
-        width: 50,
-        height: 50,
-        radius: {
-          all: 0,
-        },
-      },
-      classes: {
-        tool: {
-          'sak-rectex': true,
-          hover: true,
-        },
-        rectex: {
-          'sak-rectex__rectex': true,
-        },
-      },
-      styles: {
-        rectex: {
-        },
-      },
-    };
-    super(argToolset, Merge.mergeDeep(DEFAULT_RECTANGLEEX_CONFIG, argConfig), argPos);
-
-    this.classes.rectex = {};
-    this.styles.rectex = {};
-
-    // #TODO:
-    // Verify max radius, or just let it go, and let the user handle that right value.
-    // A q can be max height of rectangle, ie both corners added must be less than the height, but also less then the width...
-
-    const maxRadius = Math.min(this.svg.height, this.svg.width) / 2;
-    let radius = 0;
-    radius = Utils.calculateSvgDimension(this.config.position.radius.all);
-    this.svg.radiusTopLeft = +Math.min(maxRadius, Math.max(0, Utils.calculateSvgDimension(
-      this.config.position.radius.top_left || this.config.position.radius.left || this.config.position.radius.top || radius,
-    ))) || 0;
-
-    this.svg.radiusTopRight = +Math.min(maxRadius, Math.max(0, Utils.calculateSvgDimension(
-      this.config.position.radius.top_right || this.config.position.radius.right || this.config.position.radius.top || radius,
-    ))) || 0;
-
-    this.svg.radiusBottomLeft = +Math.min(maxRadius, Math.max(0, Utils.calculateSvgDimension(
-      this.config.position.radius.bottom_left || this.config.position.radius.left || this.config.position.radius.bottom || radius,
-    ))) || 0;
-
-    this.svg.radiusBottomRight = +Math.min(maxRadius, Math.max(0, Utils.calculateSvgDimension(
-      this.config.position.radius.bottom_right || this.config.position.radius.right || this.config.position.radius.bottom || radius,
-    ))) || 0;
-
-    if (this.dev.debug) console.log('RectangleToolEx constructor config, svg', this.toolId, this.config, this.svg);
-  }
-
-  /** *****************************************************************************
-  * RectangleToolEx::value()
-  *
-  */
-  set value(state) {
-    const changed = super.value = state;
-
-    return changed;
-  }
-
-  /** *****************************************************************************
-  * RectangleToolEx::_renderRectangleEx()
-  *
-  * Summary.
-  * Renders the rectangle using lines and bezier curves with precalculated coordinates and dimensions.
-  *
-  * Refs for creating the path online:
-  * - https://mavo.io/demos/svgpath/
-  *
-  */
-
-  _renderRectangleEx() {
-    this.MergeAnimationClassIfChanged();
-
-    // WIP
-    this.MergeAnimationStyleIfChanged(this.styles);
-    this.MergeAnimationStyleIfChanged();
-    if (this.config.hasOwnProperty('csnew')) {
-      this.MergeColorFromState2(this.styles.rectex, 'rectex');
-    } else {
-      this.MergeColorFromState(this.styles.rectex);
-    }
-
-    if (!this.counter) { this.counter = 0; }
-    this.counter += 1;
-
-    const svgItems = svg`
-      <g class="${classMap(this.classes.rectex)}" id="rectex-${this.toolId}">
-        <path  d="
-            M ${this.svg.x + this.svg.radiusTopLeft} ${this.svg.y}
-            h ${this.svg.width - this.svg.radiusTopLeft - this.svg.radiusTopRight}
-            q ${this.svg.radiusTopRight} 0 ${this.svg.radiusTopRight} ${this.svg.radiusTopRight}
-            v ${this.svg.height - this.svg.radiusTopRight - this.svg.radiusBottomRight}
-            q 0 ${this.svg.radiusBottomRight} -${this.svg.radiusBottomRight} ${this.svg.radiusBottomRight}
-            h -${this.svg.width - this.svg.radiusBottomRight - this.svg.radiusBottomLeft}
-            q -${this.svg.radiusBottomLeft} 0 -${this.svg.radiusBottomLeft} -${this.svg.radiusBottomLeft}
-            v -${this.svg.height - this.svg.radiusBottomLeft - this.svg.radiusTopLeft}
-            q 0 -${this.svg.radiusTopLeft} ${this.svg.radiusTopLeft} -${this.svg.radiusTopLeft}
-            "
-            counter="${this.counter}" 
-            style="${styleMap(this.styles.rectex)}"/>
-      </g>
-      `;
-    return svg`${svgItems}`;
-  }
-
-  /** *****************************************************************************
-  * RectangleToolEx::render()
-  *
-  * Summary.
-  * The render() function for this object.
-  *
-  */
-  render() {
-    return svg`
-      <g id="rectex-${this.toolId}" class="${classMap(this.classes.tool)}"
-        @click=${(e) => this.handleTapEvent(e, this.config)}>
-        ${this._renderRectangleEx()}
-      </g>
-    `;
-  }
-} // END of class
-
 /**
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  */
-
-/** ****************************************************************************
-  * EllipseTool class
-  *
-  * Summary.
-  *
-  */
-
-class EllipseTool extends BaseTool {
-  constructor(argToolset, argConfig, argPos) {
-    const DEFAULT_ELLIPSE_CONFIG = {
-      position: {
-        cx: 50,
-        cy: 50,
-        radiusx: 50,
-        radiusy: 25,
-      },
-      classes: {
-        tool: {
-          'sak-ellipse': true,
-          hover: true,
-        },
-        ellipse: {
-          'sak-ellipse__ellipse': true,
-        },
-      },
-      styles: {
-        ellipse: {
-        },
-      },
-    };
-
-    super(argToolset, Merge.mergeDeep(DEFAULT_ELLIPSE_CONFIG, argConfig), argPos);
-
-    this.svg.radiusx = Utils.calculateSvgDimension(argConfig.position.radiusx);
-    this.svg.radiusy = Utils.calculateSvgDimension(argConfig.position.radiusy);
-
-    this.classes.ellipse = {};
-    this.styles.ellipse = {};
-
-    if (this.dev.debug) console.log('EllipseTool constructor coords, dimensions', this.coords, this.dimensions, this.svg, this.config);
-  }
-
-  /** *****************************************************************************
-  * EllipseTool::_renderEllipse()
-  *
-  * Summary.
-  * Renders the ellipse using precalculated coordinates and dimensions.
-  * Only the runtime style is calculated before rendering the ellipse
-  *
-  */
-
-  _renderEllipse() {
-    this.MergeAnimationClassIfChanged();
-    this.MergeAnimationStyleIfChanged();
-    this.MergeColorFromState(this.styles.ellipse);
-
-    if (this.dev.debug) console.log('EllipseTool - renderEllipse', this.svg.cx, this.svg.cy, this.svg.radiusx, this.svg.radiusy);
-
-    return svg`
-      <ellipse class="${classMap(this.classes.ellipse)}"
-        cx="${this.svg.cx}"% cy="${this.svg.cy}"%
-        rx="${this.svg.radiusx}" ry="${this.svg.radiusy}"
-        style="${styleMap(this.styles.ellipse)}"/>
-      `;
-  }
-
-  /** *****************************************************************************
-  * EllipseTool::render()
-  *
-  * Summary.
-  * The render() function for this object.
-  *
-  */
-  render() {
-    return svg`
-      <g id="ellipse-${this.toolId}" class="${classMap(this.classes.tool)}"
-        @click=${(e) => this.handleTapEvent(e, this.config)}>
-        ${this._renderEllipse()}
-      </g>
-    `;
-  }
-} // END of class
 
 /**
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
